Changed client registration form to enter their trainer's username which the backend automatically connects to the trainer in the database as foriegnkey
This was done as there could potentially be millions of trainers and we don't want to make users search and find their trainer in the middle of the haystack. Plus it is optimal for the frontend and network bandwidth to have less unnecessary data being sent from the server to the browsers.